refactor: rename package to sqlutil
