+++ plugins/extractors/cassandra/cassandra.go
@@ -13,7 +13,7 @@ import (
 	commonv1beta1 "github.com/odpf/meteor/models/odpf/assets/common/v1beta1"
 	facetsv1beta1 "github.com/odpf/meteor/models/odpf/assets/facets/v1beta1"
 	assetsv1beta1 "github.com/odpf/meteor/models/odpf/assets/v1beta1"
-	sqlutils "github.com/odpf/meteor/plugins/utils"
+	"github.com/odpf/meteor/plugins/sqlutil"
 
 	"github.com/odpf/meteor/plugins"
 	"github.com/odpf/meteor/registry"
@@ -87,7 +87,7 @@ func (e *Extractor) Init(ctx context.Context, configMap map[string]interface{}) 
 	}
 
 	// build excluded database list
-	e.excludedKeyspaces = sqlutils.BuildBoolMap(defaultKeyspaceList)
+	e.excludedKeyspaces = sqlutil.BuildBoolMap(defaultKeyspaceList)
 
 	// connect to cassandra
 	cluster := gocql.NewCluster(e.config.Host)
