+++ plugins/extractors/cassandra/localConfig/cassandra.yaml
@@ -17,7 +17,7 @@ commitlog_segment_size_in_mb: 32
 seed_provider:
   - class_name: org.apache.cassandra.locator.SimpleSeedProvider
     parameters:
-      - seeds: "172.17.0.2"
+      - seeds: "172.17.0.3"
 concurrent_reads: 32
 concurrent_writes: 32
 concurrent_counter_writes: 32
@@ -28,14 +28,14 @@ trickle_fsync: false
 trickle_fsync_interval_in_kb: 10240
 storage_port: 7000
 ssl_storage_port: 7001
-listen_address: 172.17.0.2
-broadcast_address: 172.17.0.2
+listen_address: 172.17.0.3
+broadcast_address: 172.17.0.3
 start_native_transport: true
 native_transport_port: 9042
 start_rpc: false
 rpc_address: 0.0.0.0
 rpc_port: 9160
-broadcast_rpc_address: 172.17.0.2
+broadcast_rpc_address: 172.17.0.3
 rpc_keepalive: true
 rpc_server_type: sync
 thrift_framed_transport_size_in_mb: 15
