+++ plugins/extractors/mssql/mssql.go
@@ -16,7 +16,7 @@ import (
 	"github.com/odpf/meteor/registry"
 	"github.com/odpf/meteor/utils"
 
-	sqlutils "github.com/odpf/meteor/plugins/utils"
+	"github.com/odpf/meteor/plugins/sqlutil"
 
 	commonv1beta1 "github.com/odpf/meteor/models/odpf/assets/common/v1beta1"
 	facetsv1beta1 "github.com/odpf/meteor/models/odpf/assets/facets/v1beta1"
@@ -80,7 +80,7 @@ func (e *Extractor) Init(ctx context.Context, configMap map[string]interface{}) 
 	}
 
 	// build excluded database list
-	e.excludedDbs = sqlutils.BuildBoolMap(defaultDBList)
+	e.excludedDbs = sqlutil.BuildBoolMap(defaultDBList)
 
 	// create client
 	if e.db, err = sql.Open("mssql", e.config.ConnectionURL); err != nil {
@@ -96,7 +96,7 @@ func (e *Extractor) Extract(ctx context.Context, emit plugins.Emit) (err error) 
 	defer e.db.Close()
 	e.emit = emit
 
-	dbs, err := sqlutils.FetchDBs(e.db, e.logger, "SELECT name FROM sys.databases;")
+	dbs, err := sqlutil.FetchDBs(e.db, e.logger, "SELECT name FROM sys.databases;")
 	if err != nil {
 		return err
 	}
@@ -106,7 +106,7 @@ func (e *Extractor) Extract(ctx context.Context, emit plugins.Emit) (err error) 
 		}
 
 		tableQuery := fmt.Sprintf(`SELECT TABLE_NAME FROM %s.INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE = 'BASE TABLE';`, database)
-		tables, err := sqlutils.FetchTablesInDB(e.db, database, tableQuery)
+		tables, err := sqlutil.FetchTablesInDB(e.db, database, tableQuery)
 		if err != nil {
 			e.logger.Error("failed to get tables, skipping database", "error", err)
 			continue
